pyproject.toml
[tool.poetry]
name = "dlthub_poc"
version = "0.1.0"
description = ""
authors = ["Mutt Data"]
packages = [{include = "dlthub_poc"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
gcp-storage-emulator = "^2022.6.11"
google-api-python-client = "^2.116.0"
google-cloud-storage = "^2.14.0"
dlt = "^0.4.3"
s3fs = "^2024.2.0"
pymongo = "^4.6.1"
duckdb = "^0.9.2"
pandas = "^2.2.0"
streamlit = "^1.31.0"

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.2"
pylint = "^2.6.0"
pre-commit = "^2.7.1"
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.6.4"
pre-commit-hooks = "^3.3.0"
pytest = "^6.1.1"
shellcheck-py = "^0.7.1.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"